+++ scene/gui/graph_edit.cpp
@@ -3139,9 +3139,9 @@ GraphEdit::GraphEdit() {
 	// Allow dezooming 8 times from the default zoom level.
 	// At low zoom levels, text is unreadable due to its small size and poor filtering,
 	// but this is still useful for previewing and navigation.
-	zoom_min = (1 / Math::pow(zoom_step, 8));
+	zoom_min = (1 / Math::pow(zoom_step, static_cast<float>(8)));
 	// Allow zooming 4 times from the default zoom level.
-	zoom_max = (1 * Math::pow(zoom_step, 4));
+	zoom_max = (1 * Math::pow(zoom_step, static_cast<float>(4)));
 
 	panner.instantiate();
 	panner->set_callbacks(callable_mp(this, &GraphEdit::_pan_callback), callable_mp(this, &GraphEdit::_zoom_callback));
