+++ scene/gui/range.cpp
@@ -252,7 +252,7 @@ void Range::set_as_ratio(double p_value) {
 	if (shared->exp_ratio && get_min() >= 0) {
 		double exp_min = get_min() == 0 ? 0.0 : Math::log(get_min()) / Math::log((double)2);
 		double exp_max = Math::log(get_max()) / Math::log((double)2);
-		v = Math::pow(2, exp_min + (exp_max - exp_min) * p_value);
+		v = Math::pow(static_cast<double>(2), exp_min + (exp_max - exp_min) * p_value);
 	} else {
 		double percent = (get_max() - get_min()) * p_value;
 		if (get_step() > 0) {
