+++ scene/resources/animation.cpp
@@ -4491,7 +4491,7 @@ void Animation::_value_track_optimize(int p_idx, real_t p_allowed_velocity_err, 
 }
 
 void Animation::optimize(real_t p_allowed_velocity_err, real_t p_allowed_angular_err, int p_precision) {
-	real_t precision = Math::pow(0.1, p_precision);
+	real_t precision = Math::pow(0.1, static_cast<double>(p_precision));
 	for (int i = 0; i < tracks.size(); i++) {
 		if (track_is_compressed(i)) {
 			continue; //not possible to optimize compressed track
